+++ macros/src/parse/mod.rs
@@ -1,5 +1,6 @@
 //! Parse number from raw literals
 
+pub mod cbig;
 mod common;
 pub mod float;
 pub mod int;